pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job_insight_104"
version = "0.1.0"
description = "104 職缺市場洞察平台 (104 Job Data Insight Platform)"
readme = "README.md"
requires-python = ">=3.8, <3.12"
license = {text = "MIT"}
dependencies = [
    # Web scraping and automation
    "httpx>=0.23.0",
    "playwright>=1.30.0",
    "apscheduler>=3.9.1",
    # Data processing and analysis
    "pandas>=1.5.0",
    "numpy>=1.23.0",
    "duckdb>=0.7.0",
    "pymongo>=4.3.0",
    "pyyaml>=6.0",
    # Data visualization
    "streamlit>=1.18.0",
    "plotly>=5.13.0",
    "matplotlib>=3.6.0",
    "seaborn>=0.12.0",
    # Utilities
    "python-dotenv>=0.21.0",
    "tqdm>=4.64.0",
    "notebook>=7.3.3",
    "fire>=0.7.0",
    "boto3>=1.37.38"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
